Revert unneded use of np.float32; fix linting by returning non-np.float from Vec3.dot

+++ genesis/ext/pyrender/interaction/viewer_interaction.py
@@ -86,7 +86,7 @@ class ViewerInteraction(ViewerInteractionBase):
         right = Vec3.from_float64(mtx[:3, 0])
         up = Vec3.from_float64(mtx[:3, 1])
 
-        direction = forward + right * np.float32(x) + up * np.float32(y)
+        direction = forward + right * x + up * y
         return Ray(position, direction)
 
     def get_camera_ray(self) -> Ray:
